requirements.txt
GitPython
feedgen
markdown
